Change state to projectlist component from app
