+++ src/components/ProjectList.js
@@ -1,33 +1,74 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
+import { v4 as uuidv4 } from 'uuid'
 import addButtonShadowTop from '../assets/addButtonShadowTop.svg'
+import AddNewProjectTab from './AddNewProjectTab.js'
+import LogoHeader from './LogoHeader.js'
+import initialProjectListState from './utils/initialProjectListState.json'
 import Headline from './Headline'
-import ProjectListItem from './ProjectListItem'
 import NavigationBar from './NavigationBar'
+import ProjectListItem from './ProjectListItem'
 
-export default function ProjectList({
-  projectList,
-  updateProjectData,
-  newProject,
-}) {
+export default function ProjectList() {
+  const [projectList, setProjectList] = useState(initialProjectListState)
+  const [showAddNewProjectTab, setShowAddNewProjectTab] = useState(false)
+
+  useEffect(() => {
+    setProjectList(JSON.parse(localStorage.getItem('projects') || '[]'))
+  }, [])
+
+  useEffect(() => {
+    localStorage.setItem('projects', JSON.stringify(projectList))
+  }, [projectList])
+
   return (
-    <StyledTab>
-      <Headline
-        headlineText="Projects"
-        textColor={'var(--copper-ultralight)'}
-      />
-      <StyledProjectList>
-        {projectList.map((projectData) => (
-          <ProjectListItem
-            key={projectData.id}
-            projectData={projectData}
-            updateProjectData={updateProjectData}
+    <>
+      <LogoHeader />
+      {showAddNewProjectTab ? (
+        <AddNewProjectTab
+          onSubmit={addToProjectList}
+          onBack={() => setShowAddNewProjectTab(false)}
+        />
+      ) : (
+        <StyledTab>
+          <Headline
+            headlineText="Projects"
+            textColor={'var(--copper-ultralight)'}
           />
-        ))}
-      </StyledProjectList>
-      <NavigationBar icon={addButtonShadowTop} onClick={newProject} />
-    </StyledTab>
+          <StyledProjectList>
+            {projectList.map((projectData) => (
+              <ProjectListItem
+                key={projectData.id}
+                projectData={projectData}
+                updateProjectData={updateProjectData}
+              />
+            ))}
+          </StyledProjectList>
+          <NavigationBar
+            icon={addButtonShadowTop}
+            onClick={() => setShowAddNewProjectTab(true)}
+          />
+        </StyledTab>
+      )}
+    </>
   )
+
+  function addToProjectList(projectData, event) {
+    event.preventDefault()
+    event.target.reset()
+    event.target[0].focus()
+    projectData.id = uuidv4()
+    setProjectList([...projectList, projectData])
+  }
+
+  function updateProjectData(updatedData) {
+    const projectIndex = projectList.findIndex(
+      (project) => project.id === updatedData.id
+    )
+    let updatedProjectList = [...projectList]
+    updatedProjectList[projectIndex] = updatedData
+    setProjectList(updatedProjectList)
+  }
 }
 
 const StyledTab = styled.main`
